chromnitron/requirements_cpu.txt
pandas==2.3.1
zarr==2.18.0
numcodecs==0.14.0
tqdm==2.2.3
pyyaml==5.4.1
pyBigWig==0.3.24
git+https://github.com/tanjimin/LoRA
torch==2.5.0+cpu
torchvision==0.20.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
